+++ Assets/Scripts/SwordAssembly/SwordAssembly_Controller.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: efc462eca60114c4b90cf8652c056863
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
